requirements.txt
flask
flask-login
pymongo
python-dotenv
flask-cors
python-telegram-bot
requests